# pkg/handlers/apidocs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "sigma",
        "contact": {
            "name": "sigma",
            "url": "https://github.com/go-sigma/sigma"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0"
        },
        "version": "1.3.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/users/": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "List users with pagination",
                "parameters": [
                    {
                        "maximum": 100,
                        "minimum": 10,
                        "type": "integer",
                        "default": 10,
                        "description": "limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 1,
                        "description": "page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "sort field",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "description": "sort method",
                        "name": "method",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Username",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Response with admin",
                        "name": "without_admin",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.CommonList"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "items": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/types.UserItem"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errcode.ErrCode"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Create user",
                "parameters": [
                    {
                        "description": "User object",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.PostUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errcode.ErrCode"
                        }
                    }
                }
            }
        },
...
        "/validators/reference": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Validator"
                ],
                "summary": "Validate reference",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Reference",
                        "name": "reference",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errcode.ErrCode"
                        }
                    }
                }
            }
        },
        "/validators/regexp": {
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Validator"
                ],
                "summary": "Validate regexp",
                "parameters": [
                    {
                        "description": "Validate regexp object",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.ValidateCronRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errcode.ErrCode"
                        }
                    }
                }
            }
        },
...
        "/webhooks/{id}": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Webhook"
                ],
                "summary": "Get a webhook",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Webhook id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/types.CommonList"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "items": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/types.WebhookItem"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errcode.ErrCode"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errcode.ErrCode"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Webhook"
                ],
                "summary": "Delete a webhook",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Webhook id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errcode.ErrCode"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errcode.ErrCode"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "enums.UserRole": {
            "type": "string",
            "enum": [
                "Root",
                "Admin",
                "User",
                "Anonymous"
            ],
            "x-enum-varnames": [
                "UserRoleRoot",
                "UserRoleAdmin",
                "UserRoleUser",
                "UserRoleAnonymous"
            ]
        },
        "enums.UserStatus": {
            "type": "string",
            "enum": [
                "Active",
                "Inactive"
            ],
            "x-enum-varnames": [
                "UserStatusActive",
                "UserStatusInactive"
            ]
        },
...
        "errcode.ErrCode": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code provides a unique, string key, often capitalized with\nunderscores, to identify the error code. This value is used as the\nkeyed value when serializing api errors.",
                    "type": "string",
                    "example": "UNAUTHORIZED"
                },
                "description": {
                    "description": "Description provides a complete account of the errors purpose, suitable\nfor use in documentation.",
                    "type": "string",
                    "example": "The access controller was unable to authenticate the client. Often this will be accompanied by a Www-Authenticate HTTP response header indicating how to authenticate."
                },
                "http_status_code": {
                    "description": "HTTPStatusCode provides the http status code that is associated with\nthis error condition.",
                    "type": "integer",
                    "example": 401
                },
                "title": {
                    "description": "Title is a short, human readable description of the error condition\nincluded in API responses.",
                    "type": "string",
                    "example": "Authentication required"
                }
            }
        },
...
        "types.CommonList": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {}
                },
                "total": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "types.PostUserRequest": {
            "type": "object",
            "required": [
                "email",
                "password",
                "username"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "test@gmail.com"
                },
                "namespace_limit": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 10
                },
                "password": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 5,
                    "example": "Admin@123"
                },
                "role": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/enums.UserRole"
                        }
                    ],
                    "example": "Admin"
                },
                "username": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 2,
                    "example": "sigma"
                }
            }
        },
...
        "types.UserItem": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2006-01-02 15:04:05"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "last_login": {
                    "type": "string"
                },
                "namespace_count": {
                    "type": "integer"
                },
                "namespace_limit": {
                    "type": "integer"
                },
                "role": {
                    "$ref": "#/definitions/enums.UserRole"
                },
                "status": {
                    "$ref": "#/definitions/enums.UserStatus"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2006-01-02 15:04:05"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "types.ValidateCronRequest": {
            "type": "object",
            "required": [
                "cron"
            ],
            "properties": {
                "cron": {
                    "type": "string",
                    "example": "0 0 * * 6"
                }
            }
        },
...
        "types.WebhookItem": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2006-01-02 15:04:05"
                },
                "enable": {
                    "type": "boolean",
                    "example": true
                },
                "event_artifact": {
                    "type": "boolean",
                    "example": true
                },
                "event_daemon_task_gc": {
                    "type": "boolean",
                    "example": true
                },
                "event_member": {
                    "type": "boolean",
                    "example": true
                },
                "event_namespace": {
                    "type": "boolean",
                    "example": true
                },
                "event_repository": {
                    "type": "boolean",
                    "example": true
                },
                "event_tag": {
                    "type": "boolean",
                    "example": true
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "namespace_id": {
                    "type": "integer",
                    "example": 1
                },
                "retry_duration": {
                    "type": "integer",
                    "example": 5
                },
                "retry_times": {
                    "type": "integer",
                    "example": 3
                },
                "secret": {
                    "type": "string",
                    "example": "secret"
                },
                "ssl_verify": {
                    "type": "boolean",
                    "example": true
                },
                "updated_at": {
                    "type": "string",
                    "example": "2006-01-02 15:04:05"
                },
                "url": {
                    "type": "string",
                    "example": "http://example.com/webhook"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}